+++ tests/QA/module.tests.ps1
@@ -12,6 +12,9 @@ $SourcePath = (Get-ChildItem $ProjectPath\*\*.psd1 | Where-Object {
         $(try { Test-ModuleManifest $_.FullName -ErrorAction Stop }catch { $false }) }
     ).Directory.FullName
 
+$mut = Import-Module -Name $ProjectName -ErrorAction Stop -PassThru -Force
+$allModuleFunctions = &$mut {Get-Command -Module $args[0] -CommandType Function } $ProjectName
+
     Describe 'Changelog Management' -Tag 'Changelog' {
         It 'Changelog has been updated' -skip:(
             !([bool](Get-Command git -EA SilentlyContinue) -and
@@ -35,21 +38,15 @@ $SourcePath = (Get-ChildItem $ProjectPath\*\*.psd1 | Where-Object {
     Describe 'General module control' -Tags 'FunctionalQuality' {
 
         It 'imports without errors' {
-            { Import-Module -Name $ProjectName -Force -ErrorAction Stop } | Should Not Throw
-            Get-Module $ProjectName | Should Not BeNullOrEmpty
+            { Import-Module -Name $ProjectName -Force -ErrorAction Stop } | Should -Not -Throw
+            Get-Module $ProjectName | Should -Not -BeNullOrEmpty
         }
 
         It 'Removes without error' {
-            { Remove-Module -Name $ProjectName -ErrorAction Stop } | Should not Throw
-            Get-Module $ProjectName | Should beNullOrEmpty
+            { Remove-Module -Name $ProjectName -ErrorAction Stop } | Should -not -Throw
+            Get-Module $ProjectName | Should -beNullOrEmpty
         }
     }
-
-    #$PrivateFunctions = Get-ChildItem -Path "$ProjectPath\Private\*.ps1"
-    #$PublicFunctions =  Get-ChildItem -Path "$ProjectPath\Public\*.ps1"
-    $allModuleFunctions = @()
-    $allModuleFunctions += Get-ChildItem -Path "$SourcePath/Private/*.ps1"
-    $allModuleFunctions += Get-ChildItem -Path "$SourcePath/Public/*.ps1"
 
     if (Get-Command Invoke-ScriptAnalyzer -ErrorAction SilentlyContinue) {
         $scriptAnalyzerRules = Get-ScriptAnalyzerRule
@@ -64,27 +61,30 @@ $SourcePath = (Get-ChildItem $ProjectPath\*\*.psd1 | Where-Object {
     }
 
     foreach ($function in $allModuleFunctions) {
-        Describe "Quality for $($function.BaseName)" -Tags 'TestQuality' {
-            It "$($function.BaseName) has a unit test" {
-                Get-ChildItem "tests\" -recurse -include "$($function.BaseName).tests.ps1" | Should Not BeNullOrEmpty
+        $functionFile = Get-ChildItem -path $SourcePath -Recurse -Include "$($function.Name).ps1"
+        Describe "Quality for $($function.Name)" -Tags 'TestQuality' {
+            It "$($function.Name) has a unit test" {
+                Get-ChildItem "tests\" -recurse -include "$($function.Name).Tests.ps1" | Should Not BeNullOrEmpty
             }
 
             if ($scriptAnalyzerRules) {
-                It "Script Analyzer for $($function.BaseName)" {
-                    forEach ($scriptAnalyzerRule in $scriptAnalyzerRules) {
-                        $PSSAResult = (Invoke-ScriptAnalyzer -Path $function.FullName -IncludeRule $scriptAnalyzerRule)
-                        ($PSSAResult | Select-Object Message, Line | Out-String) | Should -BeNullOrEmpty
-                    }
+                It "Script Analyzer for $($functionFile.FullName)" {
+                    $PSSAResult = (Invoke-ScriptAnalyzer -Path $functionFile.FullName)
+                    $Report = $PSSAResult | Format-Table -AutoSize | Out-String -Width 110
+                    $PSSAResult  | Should -BeNullOrEmpty -Because `
+                        "some rule triggered.`r`n`r`n $Report"
                 }
+
             }
         }
 
-        Describe "Help for $($function.BaseName)" -Tags 'helpQuality' {
+        Describe "Help for $($function.Name)" -Tags 'helpQuality' {
             $AbstractSyntaxTree = [System.Management.Automation.Language.Parser]::
-            ParseInput((Get-Content -raw $function.FullName), [ref]$null, [ref]$null)
+            ParseInput((Get-Content -raw $functionFile.FullName), [ref]$null, [ref]$null)
             $AstSearchDelegate = { $args[0] -is [System.Management.Automation.Language.FunctionDefinitionAst] }
             $ParsedFunction = $AbstractSyntaxTree.FindAll( $AstSearchDelegate, $true ) |
-                ? Name -eq $function.BaseName
+                ? Name -eq $function.Name
+
             $FunctionHelp = $ParsedFunction.GetHelpContent()
 
             It 'Has a SYNOPSIS' {
@@ -97,8 +97,8 @@ $SourcePath = (Get-ChildItem $ProjectPath\*\*.psd1 | Where-Object {
 
             It 'Has at least 1 example' {
                 $FunctionHelp.Examples.Count | Should beGreaterThan 0
-                $FunctionHelp.Examples[0] | Should match ([regex]::Escape($function.BaseName))
-                $FunctionHelp.Examples[0].Length | Should BeGreaterThan ($function.BaseName.Length + 10)
+                $FunctionHelp.Examples[0] | Should match ([regex]::Escape($function.Name))
+                $FunctionHelp.Examples[0].Length | Should BeGreaterThan ($function.Name.Length + 10)
             }
 
             $parameters = $ParsedFunction.Body.ParamBlock.Parameters.name.VariablePath.Foreach{ $_.ToString() }